+++ src/components/10-atoms/checkbox/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.1.5
 
 - The checkbox was not able to click if no label was set (#1929)
+- The error text is always aligned correct (#1929)
 
 ## 5.1.1